:truck: domain과 global에 맞도록 파일 위치 이동 - domain 관련 파일은 모두 domain에 - SwaggerConfig는 global -> config에 위치하도록 수정함
