--- src/main/java/com/codeit/duckhu/comments/dto/request/CommentUpdateRequest.java
@@ -1,8 +1,0 @@
-package com.codeit.duckhu.comments.dto.request;
-
-import lombok.Data;
-
-@Data
-public class CommentUpdateRequest {
-  String content;
-}
