--- src/main/java/com/codeit/duckhu/comments/exception/NoCommentException.java
@@ -1,8 +1,0 @@
-package com.codeit.duckhu.comments.exception;
-
-public class NoCommentException extends RuntimeException {
-
-  public NoCommentException(String message) {
-    super(message);
-  }
-}
