--- src/main/java/com/codeit/duckhu/comments/repository/CommentRepository.java
@@ -1,11 +1,0 @@
-package com.codeit.duckhu.comments.repository;
-
-import com.codeit.duckhu.comments.domain.Comment;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, UUID> {
-
-}
