--- src/main/java/com/codeit/duckhu/config/SwaggerConfig.java
@@ -1,25 +1,0 @@
-package com.codeit.duckhu.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.servers.Server;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.List;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("DeokHuGam API")
-                        .description("DeokHuGam API 문서")
-                )
-                .servers(List.of(
-                        new Server().url("http://localhost:8080").description("Local Server")
-                ));
-    }
-}
