+++ src/main/java/com/codeit/duckhu/domain/comments/dto/request/CommentUpdateRequest.java
@@ -1,0 +1,8 @@
+package com.codeit.duckhu.domain.comments.dto.request;
+
+import lombok.Data;
+
+@Data
+public class CommentUpdateRequest {
+  String content;
+}
