+++ src/main/java/com/codeit/duckhu/domain/comments/exception/NoCommentException.java
@@ -1,0 +1,8 @@
+package com.codeit.duckhu.domain.comments.exception;
+
+public class NoCommentException extends RuntimeException {
+
+  public NoCommentException(String message) {
+    super(message);
+  }
+}
