+++ src/main/java/com/codeit/duckhu/domain/review/service/ReviewService.java
@@ -1,0 +1,9 @@
+package com.codeit.duckhu.domain.review.service;
+
+import com.codeit.duckhu.domain.review.dto.ReviewCreateRequest;
+import com.codeit.duckhu.domain.review.dto.ReviewDto;
+
+public interface ReviewService {
+  
+  ReviewDto createReview(ReviewCreateRequest createRequest);
+}
