--- src/main/java/com/codeit/duckhu/review/service/ReviewService.java
@@ -1,9 +1,0 @@
-package com.codeit.duckhu.review.service;
-
-import com.codeit.duckhu.review.dto.ReviewCreateRequest;
-import com.codeit.duckhu.review.dto.ReviewDto;
-
-public interface ReviewService {
-  
-  ReviewDto createReview(ReviewCreateRequest createRequest);
-}
